FEBioStudio: moved all 3d math to MathLib.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBioStudio@12477 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBioStudio/ModelTree.h
@@ -105,7 +105,7 @@ public:
 
 	void contextMenuEvent(QContextMenuEvent* ev) override;
 
-	int Items() const { return m_data.size(); }
+	int Items() const { return (int) m_data.size(); }
 	CModelTreeItem& GetItem(int n) { return m_data[n]; }
 
 	QTreeWidgetItem* FindItem(FSObject* o);
